# apps/rlark/docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Rlinf API
  description: Kubernetes-style API for rlinf.io resources
  version: v1alpha1
...
paths:
...
  /api/v1/rlinf.io/v1alpha1/tasks:
    get:
      tags: [Task]
      summary: List tasks
      description: List tasks resources.
      operationId: listTask
      parameters:
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/pretty'
        - $ref: '#/components/parameters/continue'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/fieldSelector'
        - $ref: '#/components/parameters/labelSelector'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          description: Unauthorized

    post:
      tags: [Task]
      summary: Create a Task
      description: Create a Task resource.
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/pretty'
        - $ref: '#/components/parameters/dryRun'
        - $ref: '#/components/parameters/fieldManager'
        - $ref: '#/components/parameters/fieldValidation'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized

    delete:
      tags: [Task]
      summary: Delete collection of tasks
      description: Delete a collection of tasks resources.
      operationId: deleteCollectionTask
      parameters:
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/pretty'
        - $ref: '#/components/parameters/continue'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/fieldSelector'
        - $ref: '#/components/parameters/labelSelector'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '401':
          description: Unauthorized

  # ─── Task Named ──────────────────────────────────────────────────
...
components:
  parameters:
...
    namespace:
      name: namespace
      in: query
      description: Namespace of the resource (required for namespaced resources)
      schema:
        type: string
    pretty:
      name: pretty
      in: query
      description: If 'true', then the output is pretty printed.
      schema:
        type: string
    continue:
      name: continue
      in: query
      description: The continue option should be set when retrieving more results from the server.
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: Limit is a maximum number of responses to return for a list call.
      schema:
        type: integer
        format: int32
    fieldSelector:
      name: fieldSelector
      in: query
      description: A selector to restrict the list of returned objects by their fields.
      schema:
        type: string
    labelSelector:
      name: labelSelector
      in: query
      description: A selector to restrict the list of returned objects by their labels.
      schema:
        type: string
    dryRun:
      name: dryRun
      in: query
      description: When present, indicates that modifications should not be persisted.
      schema:
        type: string
        enum:
          - All
    fieldManager:
      name: fieldManager
      in: query
      description: fieldManager is a name associated with the actor or entity that is making these changes.
      schema:
        type: string
    fieldValidation:
      name: fieldValidation
      in: query
      description: fieldValidation instructs the server on how to handle objects in the request that contain unknown or duplicate fields.
      schema:
        type: string
        enum:
          - Ignore
          - Warn
          - Strict
...
  schemas:
...
    ObjectMeta:
      type: object
      properties:
        name:
          type: string
          description: Name of the resource.
        generateName:
          type: string
          description: GenerateName is an optional prefix used to generate a unique name.
        namespace:
          type: string
          description: Namespace of the resource.
        selfLink:
          type: string
        uid:
          type: string
          description: UID is the unique in time and space value for this object.
        resourceVersion:
          type: string
          description: An opaque value that represents the internal version of this object.
        generation:
          type: integer
          format: int64
          description: A sequence number representing a specific generation of the desired state.
        creationTimestamp:
          type: string
          format: date-time
          description: CreationTimestamp is a timestamp representing the server time when this object was created.
        deletionTimestamp:
          type: string
          format: date-time
        deletionGracePeriodSeconds:
          type: integer
          format: int64
        labels:
          type: object
          additionalProperties:
            type: string
          description: Map of string keys and values that can be used to organize and categorize objects.
        annotations:
          type: object
          additionalProperties:
            type: string
          description: Annotations is an unstructured key value map stored with a resource.
        ownerReferences:
          type: array
          items:
            $ref: '#/components/schemas/OwnerReference'
        finalizers:
          type: array
          items:
            type: string
        managedFields:
          type: array
          items:
            type: object
    OwnerReference:
      type: object
      required:
        - apiVersion
        - kind
        - name
        - uid
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        name:
          type: string
        uid:
          type: string
        controller:
          type: boolean
        blockOwnerDeletion:
          type: boolean
...
    ListMeta:
      type: object
      properties:
        selfLink:
          type: string
        resourceVersion:
          type: string
        continue:
          type: string
          description: continue may be set if the resource set ran out of items.
        remainingItemCount:
          type: integer
          format: int64
    Condition:
      type: object
      description: Condition contains details for one aspect of the current state of this API Resource.
      required:
        - lastTransitionTime
        - message
        - reason
        - status
        - type
      properties:
        lastTransitionTime:
          type: string
          format: date-time
          description: The last time the condition transitioned from one status to another.
        message:
          type: string
          description: A human readable message indicating details about the transition.
        observedGeneration:
          type: integer
          format: int64
          description: The .metadata.generation that the condition was set based upon.
        reason:
          type: string
          description: A programmatic identifier indicating the reason for the condition's last transition.
        status:
          type: string
          enum:
            - "True"
            - "False"
            - Unknown
          description: Status of the condition, one of True, False, Unknown.
        type:
          type: string
          description: Type of condition in CamelCase or in foo.example.com/CamelCase.
    Status:
      type: object
      description: Status is a return value for calls that don't return other objects.
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          $ref: '#/components/schemas/ListMeta'
        status:
          type: string
        message:
          type: string
        reason:
          type: string
        details:
          type: object
          properties:
            name:
              type: string
            group:
              type: string
            kind:
              type: string
            uid:
              type: string
            causes:
              type: array
              items:
                type: object
                properties:
                  reason:
                    type: string
                  message:
                    type: string
                  field:
                    type: string
            retryAfterSeconds:
              type: integer
              format: int32
        code:
          type: integer
          format: int32
...
    DockerTaskSpec:
      type: object
      properties:
        containers:
          type: array
          items:
            type: object
...
    KubernetesWorkload:
      type: object
      properties:
        kind:
          type: string
        replicas:
          type: integer
        template:
          type: object
          description: PodTemplateSpec describes the data a pod should have when created from a template.
    RawTaskSpec:
      type: object
      properties:
        artifact:
          type: string
        entrypoint:
          type: array
          items:
            type: string
        environment:
          type: array
          items:
            type: object
...
    Task:
      type: object
      required:
        - apiVersion
        - kind
      properties:
        apiVersion:
          type: string
          enum:
            - rlinf.io/v1alpha1
        kind:
          type: string
          enum:
            - Task
        metadata:
          $ref: '#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/TaskSpec'
        status:
          $ref: '#/components/schemas/TaskStatus'
    TaskSpec:
      type: object
      required:
        - role
      properties:
        agentType:
          type: string
        docker:
          $ref: '#/components/schemas/DockerTaskSpec'
        kubernetes:
          $ref: '#/components/schemas/TaskKubernetesSpec'
        nodeSelector:
          type: object
          additionalProperties:
            type: string
        raw:
          $ref: '#/components/schemas/RawTaskSpec'
        role:
          type: string
        tensorBoardDir:
          type: string
    TaskKubernetesSpec:
      type: object
      properties:
        workload:
          $ref: '#/components/schemas/KubernetesWorkload'
    TaskStatus:
      type: object
      properties:
        completionTime:
          type: string
          format: date-time
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        message:
          type: string
        observedNodes:
          type: array
          items:
            type: object
        phase:
          type: string
        retryCount:
          type: integer
        startTime:
          type: string
          format: date-time
    TaskList:
      type: object
      required:
        - apiVersion
        - kind
      properties:
        apiVersion:
          type: string
          enum:
            - rlinf.io/v1alpha1
        kind:
          type: string
          enum:
            - TaskList
        metadata:
          $ref: '#/components/schemas/ListMeta'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'

    # ─── Workflow Schemas ──────────────────────────────────────────
...